# Elasticsearch output bulk response parsing