Implement simple low-level JSON parser state machine for reading elasticsearch bulk response JSON. The JSON parser requires no allocations.

benchmarks

    original (json.Umarshal):
    BenchmarkCollectPublishFailsNone-4 	  200000	      6221 ns/op	    2760 B/op	      39 allocs/op
    BenchmarkCollectPublishFailMiddle-4	  200000	      7026 ns/op	    2888 B/op	      46 allocs/op
    BenchmarkCollectPublishFailAll-4   	  200000	      8602 ns/op	    3144 B/op	      60 allocs/op

    new (manual JSON document walker):
    BenchmarkCollectPublishFailsNone-4 	 1000000	      1606 ns/op	      11 B/op	       0 allocs/op
    BenchmarkCollectPublishFailMiddle-4	 1000000	      2011 ns/op	      75 B/op	       3 allocs/op
    BenchmarkCollectPublishFailAll-4   	  500000	      2810 ns/op	     203 B/op	       9 allocs/op

Reported allocations are due to the logging system for failed items being logged (Bulk requests in tests contain only 3 elements).