@andrewkroh 2015-12-18
LGTM. Faster and fewer allocs. :+1: