@urso 2015-12-22
improved parser state machine + error handling (more strict) + some minor cleanups.

new stats:

    BenchmarkCollectPublishFailsNone-4 	 1000000	      1827 ns/op	       0 B/op	       0 allocs/op
    BenchmarkCollectPublishFailMiddle-4	 1000000	      2223 ns/op	      64 B/op	       3 allocs/op
    BenchmarkCollectPublishFailAll-4   	  500000	      3081 ns/op	     192 B/op	       9 allocs/op